Create common TitleCard component
